+++ UI/colorPicker/app.js
@@ -1,9 +1,0 @@
-App({
-  onLaunch: function () {
-    console.log('- github: https://github.com/angxuejian/moto.wxui/tree/main/UI/colorPicker')
-    console.log('- 思路与用法  详看 README.md')
-    console.log('- 有问题可以  issues  或者 QQ: 1251537708 找我哦!')
-    console.log('- v1.0')
-    console.log('\n')
-  }
-})